requirements.txt
torch
transformers
pydub
soundfile
numpy
torchvision
torchaudio
tokenizers
encodec
langid
wget
unidecode
pyopenjtalk-prebuilt
pypinyin
inflect
cn2an
jieba
eng_to_ipa
openai-whisper
matplotlib
gradio==3.41.2
nltk
sudachipy
sudachidict_core
vocos
accelerate
pyannote.audio
onnxruntime
fastapi
uvicorn[standard]
pytest
fastapi-cors
sqlalchemy
sqlalchemy.orm

